ai-dev-tools-zoomcamp/capstone_project/backend/pyproject.toml
[project]
name = "capstone-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.9",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
]
